Skips plan deletion if the job failed

+++ nautobot_golden_config/jobs.py
@@ -438,7 +438,7 @@ class DeployConfigPlans(Job):
         """Run config plan deployment process."""
         self.log_debug("Starting config plan deployment job.")
         config_deployment(self, data, commit)
-        if commit:
+        if commit and not self.failed:
             config_plan_qs = data["config_plan"]
             config_plan_qs.delete()
 
@@ -456,7 +456,8 @@ class DeployConfigPlanJobButtonReceiver(JobButtonReceiver):
         self.log_debug("Starting config plan deployment job.")
         data = {"debug": False, "config_plan": ConfigPlan.objects.filter(id=obj.id)}
         config_deployment(self, data, commit=True)
-        obj.delete()
+        if not self.failed:
+            obj.delete()
 
 
 # Conditionally allow jobs based on whether or not turned on.
